requirements.txt
asgiref==3.9.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
idna==3.11
mccabe==0.7.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
google-genai>=0.3.0

